+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 
 - [TD-4918] Implementations permissions
+- [TD-4911] Implementations download was failing for ruleless implementations
 
 ## [4.45.0] 2022-06-06
 
@@ -55,7 +56,7 @@
 
 ### Added
 
-- [TD-4089] Support for Implementations without Rules
+- [TD-4089] Support for implementations without rules
 
 ## [4.43.0] 2022-05-09